Refactor e2e test and examples
